seed_control_for_testing
lda now takes an optional seed for the PRG so that the output can be (optionally) be made
deterministic.  This is very helpful for the replication of results when testing and debugging.

When a seed is provided,  that seed is used to seed the pseudorandom generator and output will be deterministic.

When no seed is provided, as before, the time  will be used to seed the pseudorandom generator and the output will be non deterministic
(so far as the user is concerned, without knowledge of the full bits of the time at exact nanosecond the process queries the clock...)

new CLI:

lda est alpha k settings_file data random/seeded/* output_directory [PRG seed]

seeded invocation:

mpiexec.hydra -n 45 -f machinefile ./lda est 2.5 20 settings.txt model.dat random test 1812

unseeded invocation:

mpiexec.hydra -n 45 -f machinefile ./lda est 2.5 20 settings.txt model.dat random test